Handle two failure cases
The first failure case is that the hash is calculated onn the raw,
unescaped values, so ensure that we unescape for the hmac calculation.

The second failue case happens when a client requests an http resource
on behalf of another resource (ie. a <scaript> tag in an html file). In
this case, look for the proper referer.